+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-18 13:57:57
-// Commit: refactor(core): optimize data flow between domain and data layers
+// Updated at: 2026-06-19 13:36:51
+// Commit: fix(core): fix memory leak by cancelling core subscription
 
 class coreComponentState {
-  final String commitRef = '2100';
-  final String lastMessage = 'refactor(core): optimize data flow between domain and data layers';
-  final DateTime timestamp = DateTime.parse('2026-06-18T13:57:57+05:00');
+  final String commitRef = '2126';
+  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
+  final DateTime timestamp = DateTime.parse('2026-06-19T13:36:51+05:00');
 }